requirements.txt
langchain-core==0.1.32
langchain-google-genai==0.0.11
google-generativeai>=0.4.1
google-ai-generativelanguage==0.4.0
pandas==2.2.1
streamlit==1.32.2
openpyxl==3.1.2
python-dotenv==1.0.1
sentence-transformers==2.5.1
faiss-cpu==1.8.0
numpy>=1.24.0
tiktoken==0.6.0 